+++ astropy/coordinates/tests/test_representation.py
@@ -1497,3 +1497,23 @@ def test_unitphysics(unitphysics):
     assert assph.lon == obj.phi
     assert assph.lat == 80*u.deg
     assert_allclose_quantity(assph.distance, 1*u.dimensionless_unscaled)
+
+
+def test_equality_representation_differential():
+    representation_1 = CartesianRepresentation(0*u.m, 1*u.m, 2*u.m)
+    representation_2 = CartesianRepresentation(0*u.m, 1*u.m, 2*u.m)
+    assert representation_1 == representation_2
+
+    differential_1 = CartesianDifferential(0*u.m, 1*u.m, 2*u.m)
+    differential_2 = CartesianDifferential(0*u.m, 1*u.m, 2*u.m)
+    assert differential_1 == differential_2
+
+
+def test_inequality_representation_differntial():
+    representation_1 = CartesianRepresentation(0*u.m, 1*u.m, 2*u.m)
+    representation_2 = CartesianRepresentation(0*u.m, 1*u.m, 3*u.m)
+    assert representation_1 != representation_2
+
+    differential_1 = CartesianDifferential(0*u.m, 1*u.m, 2*u.m)
+    differential_2 = CartesianDifferential(0*u.m, 1*u.m, 3*u.m)
+    assert differential_1 != differential_2